Argument type fix, remove a logging statement.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@8031 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETLayoutItem.h
@@ -61,7 +61,7 @@ to be resized. */
 /** Keeps both the top margin and the height fixed but allows the bottom margin 
 to be resized. */
 };
-typedef unsigned int ETAutoresizing;
+typedef NSUInteger ETAutoresizing;
 
 /** Describes how the content looks when the layout item is resized.
